Split put lifecycle operation logic into a metadata service
